DataTable Parsing (Character, Monster)
캐릭터 정보, 몬스터 정보 테이블 데이터 파싱완료

+++ Assets/Scripts/Utils/ClassList.cs
@@ -22,13 +22,13 @@ public class CharacterData
     public int spd;
     public float spdIncrement;
     public int ats;
-    public float atsIncrement;
+    public double atsIncrement;
     public int evd;
-    public float evdIncrement;
+    public double evdIncrement;
     public int crt;
-    public float crtIncrement;
+    public double crtIncrement;
     public int crtd;
-    public float crtdIncrement;
+    public double crtdIncrement;
     public string descript;
 }
